backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
python-multipart==0.0.20
openai==1.75.0
supabase==2.15.0
python-dotenv==1.1.0 